+++ packages/calling-component-bindings/src/videoGalleryUtils.ts
@@ -1,0 +1,107 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT license.
+
+import { DominantSpeakersInfo } from '@azure/communication-calling';
+import { memoizeFnAll, toFlatCommunicationIdentifier } from '@internal/acs-ui-common';
+import { RemoteParticipantState, RemoteVideoStreamState } from '@internal/calling-stateful-client';
+import { VideoGalleryRemoteParticipant, VideoGalleryStream } from '@internal/react-components';
+import { checkIsSpeaking } from './SelectorUtils';
+
+/** @internal */
+export const _dominantSpeakersWithFlatId = (dominantSpeakers?: DominantSpeakersInfo): undefined | string[] => {
+  return dominantSpeakers?.speakersList?.map(toFlatCommunicationIdentifier);
+};
+
+/** @internal */
+export const _videoGalleryRemoteParticipantsMemo = (
+  remoteParticipants:
+    | {
+        [keys: string]: RemoteParticipantState;
+      }
+    | undefined
+): VideoGalleryRemoteParticipant[] => {
+  if (!remoteParticipants) {
+    return [];
+  }
+  return memoizedAllConvertRemoteParticipant((memoizedFn) => {
+    return Object.values(remoteParticipants)
+      .filter((participant: RemoteParticipantState) => {
+        return participant.state !== 'InLobby';
+      })
+      .map((participant: RemoteParticipantState) => {
+        return memoizedFn(
+          toFlatCommunicationIdentifier(participant.identifier),
+          participant.isMuted,
+          checkIsSpeaking(participant),
+          participant.videoStreams,
+          participant.displayName
+        );
+      });
+  });
+};
+
+const memoizedAllConvertRemoteParticipant = memoizeFnAll(
+  (
+    userId: string,
+    isMuted: boolean,
+    isSpeaking: boolean,
+    videoStreams: { [key: number]: RemoteVideoStreamState },
+    displayName?: string
+  ): VideoGalleryRemoteParticipant => {
+    return convertRemoteParticipantToVideoGalleryRemoteParticipant(
+      userId,
+      isMuted,
+      isSpeaking,
+      videoStreams,
+      displayName
+    );
+  }
+);
+
+/** @private */
+export const convertRemoteParticipantToVideoGalleryRemoteParticipant = (
+  userId: string,
+  isMuted: boolean,
+  isSpeaking: boolean,
+  videoStreams: { [key: number]: RemoteVideoStreamState },
+  displayName?: string
+): VideoGalleryRemoteParticipant => {
+  const rawVideoStreamsArray = Object.values(videoStreams);
+  let videoStream: VideoGalleryStream | undefined = undefined;
+  let screenShareStream: VideoGalleryStream | undefined = undefined;
+
+  if (rawVideoStreamsArray[0]) {
+    if (rawVideoStreamsArray[0].mediaStreamType === 'Video') {
+      videoStream = convertRemoteVideoStreamToVideoGalleryStream(rawVideoStreamsArray[0]);
+    } else {
+      screenShareStream = convertRemoteVideoStreamToVideoGalleryStream(rawVideoStreamsArray[0]);
+    }
+  }
+
+  if (rawVideoStreamsArray[1]) {
+    if (rawVideoStreamsArray[1].mediaStreamType === 'ScreenSharing') {
+      screenShareStream = convertRemoteVideoStreamToVideoGalleryStream(rawVideoStreamsArray[1]);
+    } else {
+      videoStream = convertRemoteVideoStreamToVideoGalleryStream(rawVideoStreamsArray[1]);
+    }
+  }
+
+  return {
+    userId,
+    displayName,
+    isMuted,
+    isSpeaking,
+    videoStream,
+    screenShareStream,
+    isScreenSharingOn: screenShareStream !== undefined && screenShareStream.isAvailable
+  };
+};
+
+const convertRemoteVideoStreamToVideoGalleryStream = (stream: RemoteVideoStreamState): VideoGalleryStream => {
+  return {
+    id: stream.id,
+    isAvailable: stream.isAvailable,
+    isMirrored: stream.view?.isMirrored,
+    renderElement: stream.view?.target
+  };
+};