+++ packages/communication-react/src/index.ts
@@ -196,8 +196,17 @@ export type {
   VideoTileProps,
   VideoTileStylesProps
 } from '../../react-components/src';
+/* @conditional-compile-remove-from(stable) meeting-composite Local_Camera_switcher */
+export type { LocalVideoCameraCycleButtonProps } from '../../react-components/src';
 export * from '../../react-components/src/localization/locales';
 export * from '../../react-components/src/theming';
+/* @conditional-compile-remove-from(stable) meeting-composite control-bar-split-buttons */
+export type {
+  CameraButtonContextualMenuStyles,
+  CameraButtonStyles,
+  MicrophoneButtonContextualMenuStyles,
+  MicrophoneButtonStyles
+} from '../../react-components/src';
 
 export * from '../../calling-stateful-client/src';
 export * from '../../chat-stateful-client/src';