+++ packages/react-components/src/components/styles/VideoTile.styles.ts
@@ -103,5 +103,7 @@ export const isSpeakingBorderDiv: IStyle = {
   height: '100%',
   position: 'absolute',
   top: '0',
-  left: '0'
+  left: '0',
+  // Ensure the isSpeaking element does not steal any pointer events such as onClick events
+  pointerEvents: 'none'
 };