+++ packages/react-composites/src/composites/CallComposite/components/CallControls.tsx
@@ -4,12 +4,12 @@
 import { concatStyleSets, IButtonStyles, Stack } from '@fluentui/react';
 import { _isInLobbyOrConnecting } from '@internal/calling-component-bindings';
 /* @conditional-compile-remove-from(stable): custom button injection */
-import { ControlBarButton } from '@internal/react-components';
-/* @conditional-compile-remove-from(stable): custom button injection */
-import { ControlBarButtonProps } from '@internal/react-components';
+import { ControlBarButton, ControlBarButtonProps } from '@internal/react-components';
 import {
+  BaseCustomStyles,
   CameraButton,
   ControlBar,
+  ControlBarButtonStyles,
   DevicesButton,
   EndCallButton,
   MicrophoneButton,
@@ -42,6 +42,23 @@ export type CallControlsProps = {
    * Recommended for mobile devices.
    */
   increaseFlyoutItemSize?: boolean;
+  /**
+   * Whether to use split buttons to show device selection drop-downs
+   * Used by {@link MeetingComposite}.
+   */
+  splitButtonsForDeviceSelection?: boolean;
+  /**
+   * Styles for the {@link ControlBar}.
+   */
+  controlBarStyles?: BaseCustomStyles;
+  /**
+   * Styles for all buttons except {@link EndCallButton}.
+   */
+  commonButtonStyles?: ControlBarButtonStyles;
+  /**
+   * Styles for {@link EndCallButton}.
+   */
+  endCallButtonStyles?: ControlBarButtonStyles;
 };
 
 /**
@@ -182,6 +199,8 @@ export interface CustomCallControlButtonProps extends ControlBarButtonProps {
  */
 export const CallControls = (props: CallControlsProps): JSX.Element => {
   const { callInvitationURL, onFetchParticipantMenuItems } = props;
+  const options = typeof props.options === 'boolean' ? {} : props.options;
+  const compactMode = options?.displayType === 'compact';
 
   const callStatus = useSelector(getCallStatus);
   const isLocalMicrophoneEnabled = useSelector(getLocalMicrophoneEnabled);
@@ -213,17 +232,34 @@ export const CallControls = (props: CallControlsProps): JSX.Element => {
   const devicesButtonProps = usePropsFor(DevicesButton);
   const hangUpButtonProps = usePropsFor(EndCallButton);
 
+  const commonButtonStyles = useMemo(
+    () => concatButtonBaseStyles(props.commonButtonStyles ?? {}),
+    [props.commonButtonStyles]
+  );
   const participantsButtonStyles = useMemo(
-    () => concatButtonBaseStyles(props.increaseFlyoutItemSize ? participantButtonWithIncreasedTouchTargets : {}),
-    [props.increaseFlyoutItemSize]
+    () =>
+      concatButtonBaseStyles(
+        props.increaseFlyoutItemSize ? participantButtonWithIncreasedTouchTargets : {},
+        props.commonButtonStyles ?? {}
+      ),
+    [props.increaseFlyoutItemSize, props.commonButtonStyles]
   );
-
   const devicesButtonStyles = useMemo(
-    () => concatButtonBaseStyles(props.increaseFlyoutItemSize ? devicesButtonWithIncreasedTouchTargets : {}),
-    [props.increaseFlyoutItemSize]
+    () =>
+      concatButtonBaseStyles(
+        props.increaseFlyoutItemSize ? devicesButtonWithIncreasedTouchTargets : {},
+        props.commonButtonStyles ?? {}
+      ),
+    [props.increaseFlyoutItemSize, props.commonButtonStyles]
   );
-
-  const options = typeof props.options === 'boolean' ? {} : props.options;
+  const endCallButtonStyles = useMemo(
+    () =>
+      concatStyleSets(
+        compactMode ? groupCallLeaveButtonCompressedStyle : groupCallLeaveButtonStyle,
+        props.endCallButtonStyles ?? {}
+      ),
+    [props.endCallButtonStyles]
+  );
 
   /* @conditional-compile-remove-from(stable): custom button injection */
   const customButtonProps = useMemo(() => {
@@ -238,8 +274,6 @@ export const CallControls = (props: CallControlsProps): JSX.Element => {
     return <></>;
   }
 
-  const compactMode = options?.displayType === 'compact';
-
   const microphoneButton = options?.microphoneButton !== false && (
     // tab focus on MicrophoneButton on page load
     <MicrophoneButton
@@ -247,8 +281,10 @@ export const CallControls = (props: CallControlsProps): JSX.Element => {
       data-ui-id="call-composite-microphone-button"
       {...microphoneButtonProps}
       showLabel={!compactMode}
-      styles={controlButtonBaseStyle}
+      styles={commonButtonStyles}
       {...microphoneButtonStrings}
+      /* @conditional-compile-remove-from(stable) meeting-composite control-bar-split-buttons */
+      enableDeviceSelectionMenu={props.splitButtonsForDeviceSelection}
     />
   );
 
@@ -257,7 +293,9 @@ export const CallControls = (props: CallControlsProps): JSX.Element => {
       data-ui-id="call-composite-camera-button"
       {...cameraButtonProps}
       showLabel={!compactMode}
-      styles={controlButtonBaseStyle}
+      styles={commonButtonStyles}
+      /* @conditional-compile-remove-from(stable) meeting-composite control-bar-split-buttons */
+      enableDeviceSelectionMenu={props.splitButtonsForDeviceSelection}
     />
   );
 
@@ -267,6 +305,7 @@ export const CallControls = (props: CallControlsProps): JSX.Element => {
       {...screenShareButtonProps}
       showLabel={!compactMode}
       disabled={options?.screenShareButton !== true && options?.screenShareButton?.disabled}
+      styles={commonButtonStyles}
     />
   );
 
@@ -296,7 +335,7 @@ export const CallControls = (props: CallControlsProps): JSX.Element => {
     <EndCallButton
       data-ui-id="call-composite-hangup-button"
       {...hangUpButtonProps}
-      styles={compactMode ? groupCallLeaveButtonCompressedStyle : groupCallLeaveButtonStyle}
+      styles={endCallButtonStyles}
       showLabel={!compactMode}
     />
   );
@@ -311,7 +350,7 @@ export const CallControls = (props: CallControlsProps): JSX.Element => {
             dockedBottom it has position absolute and would therefore float on top of the media gallery,
             occluding some of its content.
          */}
-        <ControlBar layout="horizontal">
+        <ControlBar layout="horizontal" styles={props.controlBarStyles}>
           {
             /* @conditional-compile-remove-from(stable): custom button injection */
             <FilteredCustomButtons customButtonProps={customButtonProps} placement={'first'} />
@@ -356,8 +395,13 @@ export const CallControls = (props: CallControlsProps): JSX.Element => {
   );
 };
 
-const concatButtonBaseStyles = (styles: IButtonStyles): IButtonStyles =>
-  concatStyleSets(controlButtonBaseStyle, styles);
+const concatButtonBaseStyles = (...styles: IButtonStyles[]): IButtonStyles => {
+  let result = controlButtonBaseStyle;
+  styles.forEach((style) => {
+    result = concatStyleSets(result, style);
+  });
+  return result;
+};
 
 /* @conditional-compile-remove-from(stable): custom button injection */
 const FilteredCustomButtons = (props: {